Now it works with all values of n. Also decode works properly now. Part 2 almost done, just left to work with stdin and stdout.

+++ hw1/include/hw1.h
@@ -11,7 +11,7 @@ int length(char *array);
 int encode();
 int decode();
 
-long stringToInt(char* string);
+int stringToInt(char* string);
 /**
 * return 0 if the strings are equal, 1 otherwise.
 */
